Setup Script For Strap
Strap will download dotfiles if the repo is named correctly. Afterwards
it will run a script/setup bash script in attempt to bootstrap the
dotfiles. Swing number one for my dotfiles. This shouldn't be much of a
problem since a modern ruby is shipped with most OSX versions these
days. Meaning that the gem install should pass with fly colors.